emacs setup

+++ make.bash
@@ -74,7 +74,9 @@ function all
     # to any hackers out there... no, I don't keep my ssh keys on disk. They are on a removable security token, so hacking
     # my dropbox account won't get you the keys
     logit "Setting up symlink from ~/Dropbox/.ssh -> ~/.ssh"
-    run "ln -s ~/Dropbox/.ssh ~/.ssh"
+    if [[ ! -s ~/Dropbox/.ssh ]]; then
+        run "ln -s ~/Dropbox/.ssh ~/.ssh"
+    fi
     logit "Setting up symlink from ~/Dropbox/.ssh -> ~/.ssh: done"
 
     logit "Copying dotfiles into $homedir"
